--- docs/database/snippets/tutorial/aspiresqlefcore/AspireSQL/Components/Pages/Home.razor
@@ -1,54 +1,0 @@
-﻿@page "/"
-@using AspireSQL.Data
-@rendermode RenderMode.InteractiveServer
-@inject TicketContext context
-
-<div class="text-center">
-    <h1 class="display-4">Request Support</h1>
-</div>
-
-<EditForm Model="@Ticket" FormName="Tickets" method="post" OnValidSubmit="@HandleValidSubmit">
-    <DataAnnotationsValidator />
-    <div class="mb-4">
-        <label>Issue Title</label>
-        <InputText class="form-control" @bind-Value="@Ticket.Title" />
-        <ValidationMessage For="() => Ticket.Title" />
-    </div>
-    <div class="mb-4">
-        <label>Issue Description</label>
-        <InputText class="form-control" @bind-Value="@Ticket.Description" />
-        <ValidationMessage For="() => Ticket.Description" />
-    </div>
-    <button class="btn btn-primary" type="submit">Submit</button>
-    <button class="btn btn-danger mx-2" type="reset" @onclick=@ClearForm>Clear</button>
-</EditForm>
-
-<table class="table table-striped">
-    @foreach (var row in rows)
-    {
-        <tr>
-            <td>@row.Id</td>
-            <td>@row.Title</td>
-            <td>@row.Description</td>
-        </tr>
-
-    }
-</table>
-
-@code {
-    private SupportTicket Ticket = new();
-    private List<SupportTicket> rows = new();
-
-    private void ClearForm() => Ticket = new();
-
-    protected override Task OnInitializedAsync()
-    {
-        return base.OnInitializedAsync();
-    }
-
-    private void HandleValidSubmit()
-    {
-        context.Tickets.Add(Ticket);
-        context.SaveChanges();
-    }
-}
