--- docs/database/snippets/tutorial/aspiresqlefcore/AspireSQL/Data/SupportTicket.cs
@@ -1,9 +1,0 @@
-﻿namespace AspireSQL.Data
-{
-    public class SupportTicket
-    {
-        public int Id { get; set; }
-        public string Title { get; set; } = String.Empty;
-        public string Description { get; set; } = String.Empty;
-    }
-}
