+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore.AppHost/Properties/launchSettings.json
