+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore.AppHost/appsettings.Development.json
