+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore.AppHost/appsettings.json
