+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore.ServiceDefaults/Extensions.cs
