+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/Components/App.razor
@@ -7,7 +7,7 @@
     <base href="/" />
     <link rel="stylesheet" href="bootstrap/bootstrap.min.css" />
     <link rel="stylesheet" href="app.css" />
-    <link rel="stylesheet" href="AspireSQL.styles.css" />
+    <link rel="stylesheet" href="AspireSQLEFCore.styles.css" />
     <link rel="icon" type="image/png" href="favicon.png" />
     <HeadOutlet />
 </head>
