+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/Components/Layout/MainLayout.razor
