+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/Components/Layout/NavMenu.razor
