+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/Components/Pages/Error.razor
