+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/Components/Pages/Weather.razor
