+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/Components/Routes.razor
