+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/Components/_Imports.razor
@@ -6,5 +6,5 @@
 @using static Microsoft.AspNetCore.Components.Web.RenderMode
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
-@using AspireSQL
-@using AspireSQL.Components
+@using AspireSQLEFCore
+@using AspireSQLEFCore.Components
