+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/Properties/launchSettings.json
