+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/SupportTicket.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace AspireSQLEFCore
+{
+    public class SupportTicket
+    {
+        public int Id { get; set; }
+        public string Title { get; set; } = string.Empty;
+        public string Description { get; set; } = string.Empty;
+    }
+}
