+++ docs/database/snippets/tutorial/aspiresqlefcore/AspireSQLEFCore/TicketContext.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.EntityFrameworkCore;
+using System.Reflection.Metadata;
+
+namespace AspireSQLEFCore
+{
+    public class TicketContext : DbContext
+    {
+        public TicketContext(DbContextOptions options) : base(options) { }
+
+        public DbSet<SupportTicket> Tickets { get; set; }
+
+    }
+}
